sass dependency added, unused images and SVGs removed, favicon updated, global styles migrated to SCSS, home page components refactored | 1
