+++ package.json
@@ -15,7 +15,8 @@
     "react-dom": "^19.1.0",
     "react-fast-marquee": "^1.6.5",
     "react-intersection-observer": "^9.16.0",
-    "react-router-dom": "^7.9.1"
+    "react-router-dom": "^7.9.1",
+    "sass": "^1.97.2"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
