--- src/app/globals.css
@@ -1,64 +1,0 @@
-@import "tailwindcss";
-
-:root {
-  --background: #ffffff;
-  --foreground: #171717;
-}
-
-@theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: "Inter", "Helvetica Neue", system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
-  --font-mono: "SFMono-Regular", "JetBrains Mono", Menlo, Consolas, "Liberation Mono", monospace;
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
-  }
-}
-
-body {
-  background-color: #ffffff;
-  color: var(--foreground);
-  font-family: var(--font-sans, Arial, Helvetica, sans-serif);
-}
-
-
-.btn{
-  @apply px-6 py-2 rounded font-semibold text-lg transition-all duration-300 flex items-center cursor-pointer;
-}
-
-.btn-primary{
-  @apply bg-black text-white border border-black;
-}
-
-.btn-outline{
-  @apply border border-black text-black;
-}
-
-.btn-primary:hover{
-  @apply bg-neutral-800;
-}
-
-.btn-outline:hover{
-  @apply bg-black text-white;
-}
-
-.custom-border{
-  @apply border-gray-400;
-}
-
-/* home page image carousel horizontal scrollbar hide */
-.scrollbar-hide {
-  overflow-x: scroll;
-  -ms-overflow-style: none; /* IE & Edge */
-  scrollbar-width: none; /* Firefox */
-}
-
-.scrollbar-hide::-webkit-scrollbar {
-  height: 0px;
-  width: 0px;
-  display: none;
-}
