+++ src/components/templateSection/DocsTemplatesSection.tsx
@@ -1,0 +1,104 @@
+import { FileText, Globe, Smile, GitBranch, Users } from "lucide-react";
+import Link from "next/link";
+
+export default function DocsTemplatesSection() {
+  return (
+    <section
+      className="w-full bg-[#f8f9fa] py-20 px-6"
+      id="docs-templates-section"
+    >
+      <div className="container mx-auto flex flex-col gap-12 bg-white p-6">
+        <div className="grid lg:grid-cols-2 gap-20 items-start">
+          {/* Left Content */}
+          <div className="space-y-6">
+            <div className="inline-flex items-center gap-2 bg-white px-4 py-2 rounded border border-gray-200">
+              <div className="w-6 h-6 bg-black rounded flex items-center justify-center">
+                <FileText className="w-4 h-4 text-white" />
+              </div>
+              <span className="font-semibold text-gray-900">
+                Knowledge base/FAQ
+              </span>
+            </div>
+
+            <div className="flex flex-col gap-1">
+              <h3 className="h3">
+                Turn Information into Actionable Knowledge base
+              </h3>
+
+              <p className="text-lg text-gray-600 leading-relaxed">
+                Create and publish beautiful, AI-optimized documentation and
+                knowledge bases that always stay up-to-date, accelerate
+                onboarding and surface in AI-powered search
+              </p>
+            </div>
+
+            <Link href="/features" target="_blank" className="btn btn-primary flex items-center justify-center gap-2 w-fit">
+              Explore more
+              <svg
+                className="w-4 h-4"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M9 5l7 7-7 7"
+                />
+              </svg>
+            </Link>
+
+            <div className="grid md:grid-cols-2 gap-4">
+              <div className="bg-white p-6 rounded-lg border border-gray-200 hover:shadow-lg transition-shadow">
+                <FileText className="w-8 h-8 text-gray-700 mb-4" />
+                <h4 className="h4">Page Analytics</h4>
+                <p className="text-sm line-clamp-3">
+                  Helps you understand how your documentation is performing. It
+                  shows important data about user activity, page health, and
+                  feedback so you can improve your content based on real
+                  insights.
+                </p>
+              </div>
+
+              <div className="bg-white p-6 rounded-lg border border-gray-200 hover:shadow-lg transition-shadow">
+                <Globe className="w-8 h-8 text-gray-700 mb-4" />
+                <h4 className="h4">Table of content in public view</h4>
+                <p className="text-sm line-clamp-3">
+                  Generates a structured outline for every public page. It helps
+                  readers quickly navigate through sections and headings,
+                  improving readability, accessibility, and overall user
+                  experience.
+                </p>
+              </div>
+
+              <div
+                className="bg-white p-6 rounded-lg border border-gray-200 hover:shadow-lg transition-shadow"
+                style={{ gridColumn: "span 2" }}
+              >
+                <Smile className="w-8 h-8 text-gray-700 mb-4" />
+                <h4 className="h4">Edit via link</h4>
+                <p className="text-sm line-clamp-3">
+                  Allows you to open a document directly in edit mode using a
+                  secure, shareable link. Instead of navigating through
+                  collections, you can jump straight to editing the specific
+                  page—making updates faster and more convenient for your team
+                </p>
+              </div>
+            </div>
+          </div>
+
+          {/* Right Content - Documentation Preview */}
+          <div className="relative w-full h-full overflow-hidden">
+            <img
+              src="/template-section-images/docstar-docs.webp"
+              alt="Documentation Interface Preview"
+              className="w-full h-full object-cover object-left"
+            />
+            <div className="absolute inset-0 bg-gradient-to-l from-white/60 via-white/30 via-40% to-transparent to-70% pointer-events-none"></div>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+}
