+++ src/scss/_button.scss
@@ -1,0 +1,23 @@
+.btn {
+  @apply px-6 py-2 rounded font-semibold text-lg transition-all duration-300 flex items-center cursor-pointer;
+}
+
+.btn-primary {
+  @apply bg-black text-white border border-black;
+}
+
+.btn-outline {
+  @apply border border-black text-black;
+}
+
+.btn-primary:hover {
+  @apply bg-neutral-800;
+}
+
+.btn-outline:hover {
+  @apply bg-black text-white;
+}
+
+.custom-border {
+  @apply border-gray-400;
+}
