+++ src/scss/_font.scss
@@ -1,0 +1,15 @@
+.h1 {
+  @apply text-3xl sm:text-5xl md:text-6xl font-medium;
+}
+
+.h2 {
+  @apply text-2xl sm:text-4xl md:text-5xl font-medium;
+}
+
+.h3 {
+  @apply text-xl sm:text-3xl md:text-4xl font-medium;
+}
+
+.h4 {
+  @apply text-lg sm:text-xl md:text-lg font-medium;
+}
