+++ src/scss/_global.scss
@@ -1,0 +1,1 @@
+@import "tailwindcss";
