+++ src/scss/_utility.scss
@@ -1,0 +1,9 @@
+/**  scrollbar classes */
+
+.scrollbar-width-none {
+  scrollbar-width: none !important;
+}
+
+.scrollbar-width-thin {
+  scrollbar-width: thin !important;
+}
